+++ app/src/main/java/pl/bk20/forest/service/StepCounterState.kt
@@ -6,6 +6,6 @@ data class StepCounterState(
     val date: LocalDate,
     val steps: Int,
     val goal: Int,
-    val distanceTravelled: Float,
+    val distanceTravelled: Double,
     val calorieBurned: Int
 )
